requirements.txt
fastapi==0.103.2
minio==7.1.17
python-dotenv==1.0.0
uvicorn==0.23.2
python-multipart==0.0.6